part4, exercises 4.15-4.23: bloglist

+++ part4/bloglist/controllers/blogs.js
@@ -1,9 +1,12 @@
 const blogsRouter = require('express').Router()
+const User = require('../models/user')
 const Blog = require('../models/blog')
 
 // get all blogs
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({})
+        .populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
@@ -11,14 +14,25 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    }
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.json(savedBlog)
 })
 
@@ -43,8 +57,28 @@ blogsRouter.put('/:id', async (request, response) => {
 
 // delete blog
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(400).json({
+            error: 'Invalid blog for deletion',
+        })
+    }
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({
+            error: 'Unauthorized to delete this blog',
+        })
+    }
 })
 
 module.exports = blogsRouter
